move Reply to seperated Pgae folder
